fix(heatmaps): address review feedback on loading changes
Guard the heatmap type lookup so an unknown type falls back to "click"
instead of throwing. URL search params, the exporter heatmap context, and
the toolbar postMessage payload all reach heatmapFilters.type without
runtime validation, and the reducer persists it, so reading .noun off an
undefined entry could take the scene down.

Move HeatmapLoadingInfo into its own file. It has three call sites and
subscribes to a logic, so it does not qualify as a private sub-piece under
the one-component-per-file rule in frontend/src/AGENTS.md.

Check mouseY against undefined rather than falsiness in both scroll depth
components, so the indicator still renders when the pointer sits on the
top row of the viewport.

Drop the heatmapValueAt comment. It explained that heatmap.js throws when
it reads a canvas created at zero height (IndexSizeError in Chromium,
NS_ERROR_FAILURE in Firefox), which is why the catch swallows rather than
crashes the scene on every mouse move.
